\name{clusterLikelihood-package}
\alias{clusterLikelihood-package}
\alias{clusterLikelihood}
\docType{package}
\title{
\packageTitle{clusterLikelihood}
}
\description{
\packageDescription{clusterLikelihood}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{clusterLikelihood}
\packageIndices{clusterLikelihood}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{clusterLikelihood}

Maintainer: \packageMaintainer{clusterLikelihood}
}
\references{
~~ Literature or other references for background information ~~
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
